requirements.txt
streamlit
pandas
scikit-learn
altair
azure-cognitiveservices-speech
openai
pydub